# internal/api/environments_handler.go
package api

import (
	"encoding/json"
	"log/slog"
	"net/http"

	"github.com/go-chi/chi/v5"
)

type EnvironmentHandler struct {
	service EnvironmentService
}

func NewEnvironmentHandler(svc EnvironmentService) *EnvironmentHandler {
	return &EnvironmentHandler{service: svc}
}

func (h *EnvironmentHandler) GetEnvironments(
	w http.ResponseWriter,
	r *http.Request,
) {
	envs, err := h.service.GetEnvironments(r.Context())
	if err != nil {
		handleServiceError(w, err, "get environments")
		return
	}

	sendSuccessResponse(w, NewEnvironmentListResponse(envs))
}

func (h *EnvironmentHandler) AddEnvironment(
	w http.ResponseWriter,
	r *http.Request,
) {
	var envReq EnvironmentRequest
	if err := json.NewDecoder(r.Body).Decode(&envReq); err != nil {
		slog.Error("error decoding request", slog.Any("error", err))
		sendErrorResponse(w, http.StatusBadRequest, "error decoding request")
		return
	}

	envModel := envReq.ToModel()

	if err := h.service.AddEnvironment(
		r.Context(),
		envModel,
		envReq.TTL,
	); err != nil {
		handleServiceError(w, err, envModel.Name)
		return
	}

	sendSuccessResponse(w, NewEnvironmentResponse(envModel))
}

func (h *EnvironmentHandler) ExtendEnvironment(
	w http.ResponseWriter,
	r *http.Request,
) {
	ctx := r.Context()
	envID := chi.URLParam(r, "id")

	var req ExtendEnvironmentRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		slog.Error("error decoding request", slog.Any("error", err))
		sendErrorResponse(w, http.StatusBadRequest, "error decoding request")
		return
	}

	env, err := h.service.ExtendEnvironment(ctx, envID, req.Period, req.Token)
	if err != nil {
		handleServiceError(w, err, envID)
		return
	}

	slog.Info("extended environment",
		slog.String("name", env.DisplayName()),
		slog.String("type", env.Type),
		slog.String("id", env.EnvID),
		slog.String("period", req.Period),
	)

	sendSuccessResponse(w, NewEnvironmentResponse(env))
}