# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: env-cleaner API
  description: |
    API for managing temporary environments lifecycle.

    Environments (Helm releases, vSphere VMs) are registered with a TTL and
    automatically deleted when they expire. Owners receive email notifications
    with a link to extend the environment lifetime if needed.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        Send `Authorization: Basic <base64(api_key)>` where `api_key` is the
        value of the `admin_api_key` server configuration option.
  schemas:
    Environment:
      type: object
      description: A managed environment registered for lifecycle tracking.
      properties:
        env_id:
          type: string
          description: Unique environment identifier.
          example: "a1b2c3d4"
        type:
          type: string
          description: Environment type.
          enum: [helm, vsphere_vm]
          example: "helm"
        name:
          type: string
          description: Environment name (Helm release name or VM name).
          example: "feature-branch-42"
        namespace:
          type: string
          description: Kubernetes namespace (Helm environments only).
          example: "dev"
        owner:
          type: string
          description: Email or username of the environment owner.
          example: "john.doe"
        delete_at:
          type: string
          description: Scheduled deletion timestamp (RFC3339).
          example: "2024-01-15 10:00:00"
      required:
        - env_id
        - type
        - name
        - owner
...
    ExtendEnvironmentRequest:
      type: object
      description: Payload for extending an environment's TTL.
      properties:
        period:
          type: string
          description: |
            Extension period added to the current deletion date.
            Supports: `Xd` (days), `Xh` (hours), `Xw` (weeks).
          example: "7d"
        token:
          type: string
          description: One-time authentication token issued to the environment owner.
          example: "abc123token"
      required:
        - period
        - token
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response payload (type depends on the endpoint).
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: integer
              description: HTTP status code.
              example: 400
            message:
              type: string
              description: Human-readable error description.
              example: "validation error: ttl is required"
paths:
...
  /api/environments/{id}/extend:
    post:
      summary: Extend environment TTL
      description: |
        Extends the scheduled deletion date of an environment by the given period.
        Uses a one-time token issued to the environment owner (delivered via email).
        No API key is required — the token authenticates the request.
      operationId: extendEnvironment
      parameters:
        - name: id
          in: path
          required: true
          description: Unique environment identifier (`env_id`).
          schema:
            type: string
          example: "a1b2c3d4"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtendEnvironmentRequest'
            example:
              period: "7d"
              token: "abc123token"
      responses:
        "200":
          description: Environment extended successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Environment'
        "400":
          description: Invalid request payload or token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Environment not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...